renaming some file s so they make sense
